fix lookup host reactive dependency issue...
I create a dependency on a lookup host changing if a helper on that
lookup host is actually used from a template. But if you change a lookup
host and the new lookup host shadows a template property or data
property, the lookup process will not rerun. This is to prevent
needlessly rerunning ALL helpers every time a lookup host (controller in
the routing use case) changes.